internal f32 calc

+++ src/data/sf2/own/preset.rs
@@ -46,8 +46,8 @@ impl Preset {
         self.prepare_max_vel_range_of_gen();
     }
 
-    pub fn get_sample(&self, key: u8, idx: usize) -> Result<i16, String> {
-        let mut sample = 0;
+    pub fn get_sample(&self, key: u8, idx: usize) -> Result<f32, String> {
+        let mut sample = 0.0;
 
         let gen_set = self.get_generator_from_key_vel(key, 64)?;
         for gen in gen_set.iter() {
@@ -59,9 +59,9 @@ impl Preset {
         Ok(sample)
     }
 
-    pub fn get_samples(&self, key: u8, start: usize, end: usize) -> Result<Vec<i16>, String> {
+    pub fn get_samples(&self, key: u8, start: usize, end: usize) -> Result<Vec<f32>, String> {
         let mut sample = Vec::new();
-        sample.resize(end - start, 0);
+        sample.resize(end - start, 0.0);
 
         let gen_set = self.get_generator_from_key_vel(key, 64)?;
         for gen in gen_set.iter() {
